+++ ratesAndConversions/src/main/kotlin/aze/talmir/task/ratesconversions/data/model/CurrencyData.kt
@@ -1,6 +1,7 @@
 package aze.talmir.task.ratesconversions.data.model
 
 import androidx.annotation.DrawableRes
+import java.math.BigDecimal
 
 /**
  * A class to hold rate conversion data which
@@ -10,7 +11,7 @@ data class CurrencyData(
     val id: Int,
     val code: String,
     val name: String,
-    val rate: Double,
+    val rate: BigDecimal,
     @DrawableRes val flag: Int,
     val isBase: Boolean = false
 )
